add some Czech translations
Part of the strings were translated blindly (mostly longer ones where
the context was apparent), the rest was done according to what was
missing when browsing through the app.

Where applicable, consistency with the existing translations was aimed
for. The translations are mostly more or less word-for-word with just a
few exceptions where the content was slightly paraphrased to better
convey the meaning.